Cache_on_disk works correctly with multiple instances

+++ arte/dataelab/cache_on_disk.py
@@ -1,11 +1,17 @@
 
 import os
 import pickle
-import numbers
 import tempfile
 from functools import wraps
+from collections import defaultdict
 
 import numpy as np
+import astropy.units as u
+
+
+class TagNotSetException(Exception):
+    '''Cache tag has not been set'''
+
 
 def cache_on_disk(f):
     '''Persistent disk cache for method return values.
@@ -40,56 +46,84 @@ def cache_on_disk(f):
      - the highest object in the hiearchy (typically the anayzer instance)
      - the new value for tmpdir or prefix
     '''
-    f.disk_cacher = DiskCacher(f)
 
+    # Bind this function to a DiskCacher constructor
+    def new_cacher():
+        return DiskCacher(f)
+
+    # Access a DiskCacher based on the instance
     @wraps(f)
     def wrapper(self, *args, **kwargs):
-        return f.disk_cacher.execute(self, *args, **kwargs)
+        return wrapper.cacher_list[self].execute(self, *args, **kwargs)
+
+    # Build a new DiskCacher whenever a new instance accesses this method
+    wrapper.cacher_list = defaultdict(new_cacher)
     return wrapper
 
 
-def set_tag(obj, tag):
-    '''Setup tag on all DiskCacher objects inside this one, recursively'''
-    for cacher in _discover_cachers(obj):
-        cacher.set_tag(tag)
+def get_disk_cacher(instance, method):
+    '''Return the DiskCacher used by *method* for instance *instance*'''
+    return method.cacher_list[instance]
 
 
-def clear_cache(obj):
-    '''Clear cache on all DiskCacher objects inside this one, recursively'''
-    for cacher in _discover_cachers(obj):
-        cacher.clear_cache()
+def set_tag(root_obj, tag):
+    '''Setup disk cache tag for root_obj and child/member objects'''
+    for instance, cacher_list in _discover_cachers(root_obj):
+        cacher_list[instance].set_tag(tag)
 
 
-def set_tmpdir(obj, tmpdir):
-    '''Set the temporary directory where cached data is stored'''
-    for cacher in _discover_cachers(obj):
-        cacher.set_tmpdir(tmpdir)
+def clear_cache(root_obj):
+    '''Clear cache for root_obj and child/member objects'''
+    for instance, cacher_list in _discover_cachers(root_obj):
+        cacher_list[instance].clear_cache()
 
 
-def set_prefix(obj, prefix):
+def set_tmpdir(root_obj, tmpdir):
+    '''Set the directory where cached data is stored'''
+    for instance, cacher_list in _discover_cachers(root_obj):
+        cacher_list[instance].set_tmpdir(tmpdir)
+
+
+def set_prefix(root_obj, prefix):
     '''Set the prefix for each tag directory'''
-    for cacher in _discover_cachers(obj):
-        cacher.set_prefix(prefix)
+    for instance, cacher_list in _discover_cachers(root_obj):
+        cacher_list[instance].set_prefix(prefix)
 
 
-def _discover_cachers(obj):
+def _discover_cachers(obj, seen=None):
     '''Generator that yields all DiskCacher objects in all child members'''
-    methods = {k: getattr(obj, k) for k in dir(obj) if callable(getattr(obj, k))}
-    members = {k: getattr(obj, k) for k in dir(obj) if not callable(getattr(obj, k))}
+
+    # This set is used to avoid infinite recursion
+    # when objects have circular references
+    if seen is None:
+        seen = set()
+
+    # Some packages like astropy.units define attributes
+    # that just raise exceptions if they are accessed (!)
+    valid_attrs = {}
+    for name in dir(obj):
+        try:
+            valid_attrs[name] = getattr(obj, name)
+        except AttributeError:
+            pass
+
+    methods = {k: v for k, v in valid_attrs.items() if callable(v)}
+    members = {k: v for k, v in valid_attrs.items() if not callable(v)}
 
     # Examine both methods and members, because DiskCacher replaces a method
-    for _, method in methods.items():
-        if hasattr(method, 'disk_cacher'):
-            yield method.disk_cacher
+    for name, method in methods.items():
+        seen.update([type(method)])
+        if hasattr(method, 'cacher_list'):
+            yield obj, method.cacher_list
 
     for name, member in members.items():
-        if hasattr(member, 'disk_cacher'):
-            yield member.disk_cacher
+        if hasattr(member, 'cacher_list'):
+            yield obj, member.cacher_list
 
         # Recurse into all child members except for special variables.
-        # For some reason, numeric types have infinite recursion
-        if not name.startswith('__') and not isinstance(member, numbers.Number):
-            yield from _discover_cachers(member)
+        if (not name.startswith('__') and not type(member) in seen):
+            seen.update([type(member)])
+            yield from _discover_cachers(member, seen=seen)
 
 #    What happens with properties?
 #
@@ -100,6 +134,7 @@ def _discover_cachers(obj):
 
 
 class DiskCacher():
+    '''Class implementing a persistent disk cache'''
 
     def __init__(self, f):
         self._tag = None
@@ -124,7 +159,8 @@ class DiskCacher():
     def clear_cache(self):
         '''Remove cache from disk and memory'''
         self._data = None
-        self._delete_from_disk()
+        if self._tag:
+            self._delete_from_disk()
 
     def execute(self, *args, **kwargs):
         '''Cache lookup'''
@@ -143,48 +179,57 @@ class DiskCacher():
         else:
             return self._original_function(*args, **kwargs)
 
-    def _fullpath_no_extension(self):
+    def fullpath_no_extension(self):
         '''Full path of cache file on disk, without extension'''
+        if self._tag is None:
+            raise TagNotSetException('Disk cache tag has not been set')
         return os.path.join(self._tmpdir, self._prefix + self._tag, self._funcid)
 
-    def _file_on_disk(self):
+    def file_on_disk(self):
         '''Returns the cache file path, raises FileNotFoundError if not found'''
         extensions = ['.pickle', '.npy']
         for ext in extensions:
-            path = self._fullpath_no_extension() + ext
+            path = self.fullpath_no_extension() + ext
             if os.path.exists(path):
                 return path
         raise FileNotFoundError
 
     def _save_to_disk(self):
-        fullpath = self._fullpath_no_extension()
+        fullpath = self.fullpath_no_extension()
         try:
             os.mkdir(os.path.join(self._tmpdir, self._prefix+self._tag))
         except FileExistsError:
             pass
         if isinstance(self._data, np.ndarray):
-            np.save(fullpath+'.npy', self._data)
+            if isinstance(self._data, u.Quantity):
+                np.save(fullpath+'.npy', self._data.value)
+                with open(fullpath+'.npy.unit.pickle', 'wb') as f:
+                    pickle.dump(self._data.unit, f)
+            else:
+                np.save(fullpath+'.npy', self._data)
         else:
             with open(fullpath+'.pickle', 'wb') as f:
                 pickle.dump(self._data, f)
 
     def _load_from_disk(self):
-        path = self._file_on_disk()
+        path = self.file_on_disk()
         if path.endswith('.npy'):
-            return np.load(path)
+            data = np.load(path)
+            print('Loading from disk')
+            if os.path.exists(path+'.unit.pickle'):
+                with open(path+'.unit.pickle', 'rb') as f:
+                    unit = pickle.load(f)
+                data = data * unit
+            return data
         else:
             with open(path, 'rb') as f:
                 return pickle.load(f)
 
     def _delete_from_disk(self):
-        try:
-            path = self._file_on_disk()
-            os.unlink(path)
-        except FileNotFoundError:
-            pass
+        extensions = ['.pickle', '.npy', 'unit.pickle']
+        for ext in extensions:
+            path = self.fullpath_no_extension() + ext
+            if os.path.exists(path):
+                os.unlink(path)
 
-#
-
-
-
-
+# __oOo__
